requirements.txt
tensorflow
gunicorn
pillow
numpy
flask